feat: Set reclaim policy to delete for pvc
feat: Set reclaim policy to delete for pvc
After running runboat for quite some time, we have seen a lot of PV
(persistent volumes) that are not attached (Released or Available), but
the reclaim policy is set to Retain.
This makes the cluster keep a lot of persistent volumes that we don't
use, and we can end up hitting the maximum of volumes quite rapidly,
since each volumes is 1GB sized.

To fix this issue, we should change the policy to Delete instead, by
modifying the pvc.yaml file in runboat.
The whole documentation is available [here](https://kubernetes.io/docs/concepts/storage/persistent-volumes/#reclaim-policy)